Rename Exploration.Maps to Exploration.Charts
Chart is defined as 'Special-purpose map designed for navigation or to present specific data or information.'

The reason for this change is a naming conflict between v2/maps and v2/continents. The maps endpoint provides only a simplified view of each map. The continents endpoint provides the same maps with additional information. (Waypoints, landmarks, vistas and more.)