Improve how dev container is built
The devcontainer.json is now configured to build an image from the Dockerfile instead of pulling a pre-built image. Importantly, the pre-built image is still used as a cache. So while pulling the image can still take a while as before, building the source Dockerfile is now super fast.

This approach enables testing changes to the Dockerfile without making changes to devcontainer.json.