Improve querying by ids
- GetByIds now always return a Replica<HashSet<T>>
- Previously some GetByIds were implemented as IAsyncEnumerable
- GetBulk methods are added that return the IAsyncEnumerable

The GetBulk methods are only added for
endpoints that do not support ids=all (at least for now)

- v2/items
- v2/recipes
- v2/commerce/prices
- v2/commerce/listings

GetBulk will take a list of ids and partition them into chunks of 200,
then fetch those chunks in parallel. You can optionally configure the
chunk size and degree of parallelism. As a user, you do not really
perceive that this chunking is happening in the background. It is an
effective substitute for ids=all. Perhaps the only downside is you
cannot access the response headers of individial chunks.