Improve bulk query throughput
Previously, the request limiter would wait for a response to be fully processed by calling code before allowing the next request to begin. Now, the next request is allowed to begin immediately after a previous request ended, before the response is yielded to calling code. This way, the performance of the calling code is not a concern for network throughput.